[refactor]: 파일 위치 변경
