+++ backend/src/main/java/com/ddbb/dingdong/presentation/endpoint/auth/exchanges/CheckEmailDto.java
@@ -1,4 +1,4 @@
-package com.ddbb.dingdong.presentation.endpoint.auth.dto;
+package com.ddbb.dingdong.presentation.endpoint.auth.exchanges;
 
 import lombok.Data;
